+++ backend/src/main/java/edu/ntnu/idatt2106_2023_06/backend/dto/stat/StatAddItemToFridgeDTO.java
@@ -1,6 +1,6 @@
 package edu.ntnu.idatt2106_2023_06.backend.dto.stat;
 
-public record StatAddItemToFridgeDTO(Integer price,
+public record StatAddItemToFridgeDTO(Double price,
                                      Integer quantity,
                                      String itemName,
                                      String storeName,
